+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 `fixnum` adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.2 - 2020-12-09
+### Added
+- [`parity-scale-codec`](https://docs.rs/parity-scale-codec) support under the `parity` feature
+  (`Encode` and `Decode` impls; @quasiyoke).
+- `fixnum!` macro for compile-time-checked fixed-point "literals".
+- `.as_str()` was implemented for errors (@quasiyoke).
+
+### Changed
+- Added `$crate` prefix for `impl_op` macro (@quasiyoke).
+
 ## 0.2.1 - 2020-12-04
 ### Changed
 - Docs' links were fixed (@quasiyoke).